+++ src/constants.ts
@@ -8,3 +8,7 @@ export const MAXIMUM_RECURSION_DEPTH = 1000
 export const COMMENT_REG_EXP = /\/\*.*\*\//g
 export const PARAM_REG_EXP = /[0-9a-zA-Z_]*:/g
 export const OPT_PARAM_REG_EXP = /[0-9a-zA-Z_]*\?:/g
+
+export const APP_NAME = 'ts-for-gir'
+export const APP_USAGE = 'Typescript .d.ts generator from GIR for gjs or node-gtk'
+export const APP_SOURCE = 'https://github.com/sammydre/ts-for-gjs'