+++ src/types/generate-config.ts
@@ -1,4 +1,4 @@
-import { BuildType, Environment } from '.'
+import { BuildType, Environment, ModuleType } from '.'
 
 /**
  * Type for currently used config e.g. in GirModule
@@ -10,14 +10,16 @@ export interface GenerateConfig {
     outdir: string | null
     /** Definitions generation type */
     buildType?: BuildType
+    /** Module type, can be CommonJS or ESM */
+    moduleType?: ModuleType
     /** GIR directories */
     girDirectories: string[]
     /** Switch on/off the verbose mode */
     verbose: boolean
     /** prettifies the generated .d.ts files */
     pretty: boolean
-    /** Export all symbols for each module as a single entity using ES6 export default */
-    exportDefault: boolean
+    /** Export all symbols for each module as a namespace */
+    exportNamespace: boolean
     /** Do not generate documentation comments */
     noComments: boolean
 }