+++ test/ava/test.ts
@@ -25,7 +25,7 @@ const config: GenerateConfig = {
     pretty: false,
     verbose: false,
     buildType: 'lib',
-    exportDefault: false,
+    exportNamespace: false,
 }
 
 test('enumeration', (t) => {